[aws-datastore] Move model concerns into model subpackage
The various implementations of ModelProvider (and the
ModelProviderLocator) are concerns of the data modeling aspect of the
DataStore.

These components are moved into the newly created
com.amplifyframework.datastore.model subpackage, which deals with such
concerns.
